requirements.txt
streamlit>=1.37
requests>=2.31
pillow>=10.0
